Move & refactor BaseColorsSection contents
