+++ src/components/layout/index.ts
@@ -1,4 +1,3 @@
-export * from './BaseColorsSection';
 export * from './Header';
 export * from './OptionsDrawer';
 export * from './ThemeStyle';
